Move helpers into ethereum/helpers
