+++ ethereum/contracts/JugHelper.ts
@@ -1,6 +1,6 @@
 import { Jug__factory } from 'generated/types';
 
-import { toFixedNumber, UnitFormats } from '../math';
+import { toFixedNumber, UnitFormats } from '../helpers/math';
 
 import type EthereumProvider from '../EthereumProvider';
 import type IlkType from 'ethereum/IlkType';
