+++ ethereum/contracts/SpotHelper.ts
@@ -1,6 +1,6 @@
 import { Spotter__factory } from 'generated/types';
 
-import { toFixedNumber, UnitFormats } from '../math';
+import { toFixedNumber, UnitFormats } from '../helpers/math';
 
 import type EthereumProvider from '../EthereumProvider';
 import type IlkType from '../IlkType';
